Updated the build system and added logging source
Updated the build system to allow for source files to be placed in
subdirectories. A hierarchy of SConscript files will return a list
of objects that are to be built into the engine binary. Also added
a temporary source file that will enventually be used to handle
all of the logging and added log4cplus into the linked libraries.