Update ContentManager API
Made the contentmanager class easier to use. Seperate maps will be
maintained for each file type and there will be methods for returning
each filetype. This should have the benefit of improving speed since
we'll be looking through fewer objects each lookup and also being
much neater to work with.